# Added :hover to .mod-actions-menu-btn