## What type of PR is this? (check all applicable)

- [ ] Refactor
- [ ] Feature
- [x] Bug Fix
- [ ] Optimization
- [ ] Documentation Update

## Description
Just added `:hover` to `.mod-actions-menu-btn`


```diff
.mod-actions-menu-btn {
  position: fixed;
  bottom: var(--su-7);
  right: var(--su-7);
  z-index: 201;
  cursor: pointer;
  transition: 0.3s all ease;
  border-radius: 50px;
  
+  &:hover {
+    z-index: 201;
+  }

[...]
```

## Related Tickets & Documents
/
## Mobile & Desktop Screenshots/Recordings (if there are UI changes)

| **Old** | **New** |
|-|-|
| ![image](https://user-images.githubusercontent.com/14293805/84569916-37b62f00-ad8a-11ea-8b55-d95214217cc1.png)  | ![image](https://user-images.githubusercontent.com/14293805/84569921-4d2b5900-ad8a-11ea-9a42-4d0041e34171.png)|
## Added tests?

- [ ] yes
- [x] no, because they aren't needed
- [ ] no, because I need help

## Added to documentation?

- [ ] docs.dev.to
- [ ] readme
- [x] no documentation needed